+++ src/main/java/cz/thepatik/welcomeplugin/commands/CommandManager.java
@@ -7,7 +7,6 @@ import net.md_5.bungee.api.chat.ClickEvent;
 import net.md_5.bungee.api.chat.ComponentBuilder;
 import net.md_5.bungee.api.chat.HoverEvent;
 import net.md_5.bungee.api.chat.TextComponent;
-import org.bukkit.Bukkit;
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandExecutor;
 import org.bukkit.command.CommandSender;
@@ -15,7 +14,6 @@ import org.bukkit.command.TabExecutor;
 import org.bukkit.entity.Player;
 
 import java.util.ArrayList;
-import java.util.Objects;
 
 public class CommandManager implements CommandExecutor, TabExecutor {
     WelcomePlugin welcomePlugin;
@@ -24,27 +22,35 @@ public class CommandManager implements CommandExecutor, TabExecutor {
     }
     private ArrayList<SubCommand> subcommands = new ArrayList<>();
 
+    // Register all commands
     public CommandManager(){
         subcommands.add(new UpdateCommand());
         subcommands.add(new VersionCommand());
         subcommands.add(new HelpCommand());
         subcommands.add(new ShowCreditsToCommand(WelcomePlugin.getPlugin()));
         subcommands.add(new PlayedTimeCommand());
+        subcommands.add(new PlayerJoinsCommand());
     }
 
     @Override
     public boolean onCommand(CommandSender sender, Command cmd, String label, String[] args) {
         String command = cmd.toString();
         if (sender instanceof Player){
+            // Run command if sender is a player
+
             Player p = (Player) sender;
 
             if (args.length > 0){
+                // Run command
+
                 for (int i = 0; i < getSubcommands().size(); i++){
                     if (args[0].equalsIgnoreCase(getSubcommands().get(i).getName())){
                         getSubcommands().get(i).perform(p, args);
                     }
                 }
             } else if (args.length == 0){
+                // Run when no subcommand
+
                 TextComponent welcomeHelpMessage = new TextComponent("/welcome help");
                 welcomeHelpMessage.setColor(ChatColor.RED);
                 welcomeHelpMessage.setBold(false);
@@ -59,25 +65,47 @@ public class CommandManager implements CommandExecutor, TabExecutor {
                sender.spigot().sendMessage(welcomeHelpMessageFirst, welcomeHelpMessage);
             }
         }
+
         return true;
+
     }
 
     public ArrayList<SubCommand> getSubcommands(){
         return subcommands;
     }
 
+    // Tab complete logic
     @Override
     public ArrayList<String> onTabComplete(CommandSender sender, Command cmd, String alias, String[] args){
-        ArrayList<String> emptyList = new ArrayList<>();
+        ArrayList<String> completions = new ArrayList<>();
 
         if (args.length == 1){
-            ArrayList<String> subcommandsArguments = new ArrayList<>();
-
+            // Find subcommand
             for (int i = 0; i < getSubcommands().size(); i++){
-                subcommandsArguments.add(getSubcommands().get(i).getName());
+                completions.add(getSubcommands().get(i).getName());
             }
-            return subcommandsArguments;
+        } else if (args.length >= 2) {
+            // Find corresponding subcommand
+            SubCommand subCommand = null;
+            for (SubCommand sc : subcommands) {
+                if (sc.getName().equalsIgnoreCase(args[0])) {
+                    subCommand = sc;
+                    break;
+                }
+            }
+
+            if (subCommand != null) {
+                // Autocomplete suggestions for subcommand arguments
+                String[] subArgs = new String[args.length - 1];
+                System.arraycopy(args, 1, subArgs, 0, subArgs.length);
+                completions.addAll(subCommand.tabComplete(subArgs));
+            }
         }
-        return emptyList;
+
+        // Filter and return autocomplete suggestions based on what the user has typed
+        String currentArg = args[args.length - 1].toLowerCase();
+        completions.removeIf(suggestion -> !suggestion.toLowerCase().startsWith(currentArg));
+
+        return completions;
     }
 }