+++ src/main/java/cz/thepatik/welcomeplugin/commands/subcommands/PlayedTimeCommand.java
@@ -3,6 +3,7 @@ package cz.thepatik.welcomeplugin.commands.subcommands;
 import cz.thepatik.welcomeplugin.WelcomePlugin;
 import cz.thepatik.welcomeplugin.commands.SubCommand;
 import cz.thepatik.welcomeplugin.database.Database;
+import me.clip.placeholderapi.PlaceholderAPI;
 import org.bukkit.ChatColor;
 import org.bukkit.entity.Player;
 
@@ -11,6 +12,8 @@ import java.math.RoundingMode;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
+import java.util.Collections;
+import java.util.List;
 
 public class PlayedTimeCommand extends SubCommand {
     @Override
@@ -34,28 +37,19 @@ public class PlayedTimeCommand extends SubCommand {
     }
 
     @Override
+    public boolean hasArguments() {
+        return true;
+    }
+    @Override
+    public List<String> tabComplete(String[] args){
+        return Collections.emptyList();
+    }
+    @Override
     public void perform(Player player, String[] args) {
-        Database database = WelcomePlugin.database;
-
-        try{
-            PreparedStatement preparedStatement = database.connection.prepareStatement("SELECT PlayTime FROM PlayerData WHERE PlayerUUID = ?");
-            preparedStatement.setString(1, player.getUniqueId().toString());
-            ResultSet resultSet = preparedStatement.executeQuery();
-            int playTime = resultSet.getInt("PlayTime");
-
-            double playTimeMinutes = (double) playTime /60;
-
-            BigDecimal playTimeDecimal = BigDecimal.valueOf(playTimeMinutes);
-            playTimeDecimal = playTimeDecimal.setScale(1, RoundingMode.DOWN);
-
-            if (player.hasPermission(getPermissions())){
-                player.sendMessage(ChatColor.GREEN + "You are playing on the server for " + playTimeDecimal);
-            }else {
+       if (player.hasPermission(getPermissions())){
+                player.sendMessage(ChatColor.GREEN + "You are playing on the server for " + ChatColor.GOLD + PlaceholderAPI.setPlaceholders(player.getPlayer(), "%WelcomePlugin_played_time%"));
+       }else {
                 player.sendMessage(ChatColor.RED + "You do not have permissions!");
-            }
-
-        } catch (SQLException e) {
-            throw new RuntimeException(e);
-        }
+       }
     }
 }