+++ src/main/java/cz/thepatik/welcomeplugin/commands/subcommands/ShowCreditsToCommand.java
@@ -6,7 +6,8 @@ import net.md_5.bungee.api.ChatColor;
 import org.bukkit.configuration.ConfigurationSection;
 import org.bukkit.entity.Player;
 
-import javax.security.auth.login.Configuration;
+import java.util.ArrayList;
+import java.util.List;
 
 public class ShowCreditsToCommand extends SubCommand {
 
@@ -37,6 +38,23 @@ public class ShowCreditsToCommand extends SubCommand {
     }
 
     @Override
+    public boolean hasArguments() {
+        return true;
+    }
+
+    @Override
+    public List<String> tabComplete(String[] args) {
+        List<String> completions = new ArrayList<>();
+
+        if (args.length == 1){
+            completions.add("newcomers");
+            completions.add("everyone");
+            completions.add("nobody");
+        }
+        return completions;
+    }
+
+    @Override
     public void perform(Player player, String[] args) {
         ConfigurationSection cs = plugin.getConfig().getConfigurationSection("settings");
 
@@ -44,7 +62,7 @@ public class ShowCreditsToCommand extends SubCommand {
             if (args.length == 1){
                 String toWho = cs.getString("show-credits");
                 player.sendMessage(ChatColor.GREEN + "Now showing credits to: " + ChatColor.DARK_GREEN + toWho);
-                player.sendMessage(ChatColor.RED + "To edit you must specify to who -" + ChatColor.DARK_GREEN + " newcomers" + ChatColor.RED + " or" + ChatColor.DARK_GREEN + " everyone");
+                player.sendMessage(ChatColor.RED + "To edit you must specify to who -" + ChatColor.DARK_GREEN + " newcomers" + ChatColor.RED + " or" + ChatColor.DARK_GREEN + " everyone" + ChatColor.RED + " or" + ChatColor.DARK_GREEN + " nobody");
             } else if (args.length == 2){
                 String toWho = args[1];