+++ src/main/java/cz/thepatik/welcomeplugin/commands/subcommands/UpdateCommand.java
@@ -8,8 +8,11 @@ import net.md_5.bungee.api.chat.TextComponent;
 import org.bukkit.ChatColor;
 import org.bukkit.entity.Player;
 
-import static cz.thepatik.welcomeplugin.VersionCheck.getCurrentOnlineVersion;
-import static cz.thepatik.welcomeplugin.VersionCheck.pluginVersion;
+import java.util.Collections;
+import java.util.List;
+
+import static cz.thepatik.welcomeplugin.utils.VersionCheck.getCurrentOnlineVersion;
+import static cz.thepatik.welcomeplugin.utils.VersionCheck.pluginVersion;
 import static java.lang.Double.parseDouble;
 
 public class UpdateCommand extends SubCommand {
@@ -32,6 +35,16 @@ public class UpdateCommand extends SubCommand {
     @Override
     public String getSyntax() {
         return "/welcome update";
+    }
+
+    @Override
+    public boolean hasArguments() {
+        return false;
+    }
+
+    @Override
+    public List<String> tabComplete(String[] args) {
+        return Collections.emptyList();
     }
 
     @Override