+++ src/main/java/cz/thepatik/welcomeplugin/tasks/PlayTimeTask.java
@@ -1,6 +1,6 @@
 package cz.thepatik.welcomeplugin.tasks;
 
-import cz.thepatik.welcomeplugin.PlayerListener;
+import cz.thepatik.welcomeplugin.utils.handlers.PlayerHandler;
 import cz.thepatik.welcomeplugin.WelcomePlugin;
 import org.bukkit.entity.Player;
 import org.bukkit.scheduler.BukkitRunnable;
@@ -37,10 +37,10 @@ public class PlayTimeTask extends BukkitRunnable {
         }
     }
 
-    PlayerListener playerListener;
+    PlayerHandler playerHandler;
 
-    public PlayTimeTask(PlayerListener playerListener, Player player){
-        this.playerListener = playerListener;
+    public PlayTimeTask(PlayerHandler playerHandler, Player player){
+        this.playerHandler = playerHandler;
         p = player;
     }